+++ hw12_13_14_15_calendar/internal/server/http/middleware.go
@@ -1,7 +1,9 @@
 package internalhttp
 
-// func loggingMiddleware(next http.Handler) http.Handler {
-// 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-// 		// TODO
-// 	})
-// }
+import "net/http"
+
+func loggingMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		// TODO
+	})
+}
